// Alize.Platform.Client/src/app/components/models/company.model.ts
export class Company {
    id: string;
    name: string;
    description: string;
    isActive: boolean;
    activity: string;
    businessName: string;
    cif: string;
    comments: string;
    language: string;
    phoneNumber: string;
    email: string;
    web: string;
    contactName: string;
    logo: string;
    imageTypeMime: string;
    address: string;
    zip: string;
    city: string;
    province: string;
    country: string;
    googleMapsUrl: string;
}